cronjobs/requirements.txt
praw
psaw
python-dotenv
typer
coloredlogs
python-dateutil
pydantic
cysystemd
